refactor(synapse): isolate pipeline metrics and extract execution logic
- Move PipelineMetrics class to a dedicated utility file for better SoC
- Extract layer execution loop from process() to private _executeLayers method
- Strengthen engine test assertions to verify exact layer loading count
